requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
openai  
python-dotenv
python-multipart
pydantic
openpyxl